+++ src/app/app.module.ts
@@ -4,6 +4,7 @@ import { ContentModule } from '../modules/content/content.module';
 import { DatabaseModule } from '../modules/database/database.module';
 import { EventModule } from '../modules/event';
 import { FlowModule } from '../modules/flow/flow.module';
+import { HealthCheckModule } from '../modules/health-check/health-check.module';
 import { HttpModule } from '../modules/http/http.module';
 import { IdentityModule } from '../modules/identity/identity.module';
 import { LambdaModule } from '../modules/lambda/lambda.module';
@@ -28,6 +29,7 @@ import { LogLambda } from './logger/log.lambda';
     DatabaseModule,
     EventModule,
     FlowModule,
+    HealthCheckModule,
     HttpModule,
     IdentityModule,
     LambdaModule,