+++ src/modules/page/service/page.service.ts
@@ -1,8 +1,5 @@
-import { readFile } from 'fs/promises';
 import { Model } from 'objection';
-import { join } from 'path';
 import { ILogger, Inject, Logger, Service } from '../../../app/container';
-import { SEED_DIR } from '../../../app/globals';
 import { SchemaService } from '../../schema/service/schema.service';
 import { IPage } from '../interface/page.interface';
 
@@ -10,8 +7,6 @@ type PageModel = IPage & Model;
 
 @Service()
 export class PageService {
-  protected isSeeded: boolean = false;
-
   constructor(
     @Logger()
     readonly logger: ILogger,
@@ -20,12 +15,6 @@ export class PageService {
   ) {}
 
   async loadRoutes(): Promise<IPage[]> {
-    if (!this.isSeeded) {
-      await this.seed();
-
-      this.isSeeded = true;
-    }
-
     const model = this.schema.getModel<PageModel>('main', 'Page');
     const pages = await model
       .query()
@@ -46,19 +35,5 @@ export class PageService {
     </html>`;
 
     return html;
-  }
-
-  async seed() {
-    const model = this.schema.getModel('main', 'Page');
-
-    // ALready exists
-    if (await model.query().findById('1e1b9598-f8b8-4487-9b7b-166a363e8ce8')) {
-      return;
-    }
-
-    const landing = await readFile(join(SEED_DIR, 'landing.page.json'));
-
-    await model.query().insert(JSON.parse(landing.toString()));
-    this.logger.info('Pages seeded');
   }
 }